+++ node/clients/Organizations.ts
@@ -12,15 +12,15 @@ import type {
 const getPersistedQuery = () => {
   return {
     persistedQuery: {
-      provider: 'vtex.b2b-organizations-graphql@0.x',
-      sender: 'vtex.storefront-permissions@1.x',
+      provider: 'vtex.b2b-organizations-graphql@1.x',
+      sender: 'vtex.storefront-permissions@2.x',
     },
   }
 }
 
 export class OrganizationsGraphQLClient extends AppGraphQLClient {
   constructor(ctx: IOContext, options?: InstanceOptions) {
-    super('vtex.b2b-organizations-graphql@0.x', ctx, options)
+    super('vtex.b2b-organizations-graphql@1.x', ctx, options)
   }
 
   public getOrganizationById = async (orgId: string): Promise<unknown> => {
@@ -82,7 +82,11 @@ export class OrganizationsGraphQLClient extends AppGraphQLClient {
         page,
         pageSize,
       },
-    }) as Promise<{ data: { getOrganizationsPaginatedByEmail:  GetOrganizationsPaginatedByEmailResponse }}>
+    }) as Promise<{
+      data: {
+        getOrganizationsPaginatedByEmail: GetOrganizationsPaginatedByEmailResponse
+      }
+    }>
   }
 
   private query = async (param: {